requirements.txt
google-search-results
duckduckgo-search
langchain-community
langchain-google-genai
langchainhub
langchain-groq
wikipedia
newsapi-python
langchain-cohere
tavily-python
pandas
python-dotenv
fastapi 
uvicorn 
python-multipart
streamlit
streamlit_gsheets
